--- HomeWork13/CustomModule/Constants.swift
@@ -1,72 +1,0 @@
-//
-//  Constants.swift
-//  HomeWork13
-//
-//  Created by Игорь Николаев on 12.01.2023.
-//
-
-import UIKit
-
-//MARK: - Constants Model
-
-// Strings
-
-enum ConstantsName {
-    static let user = "Игорь Николаев"
-    static let infoUser = "Apple ID, iCloude, контент\n и покупки"
-    static let airplaneMode = "Авиарежим"
-    static let wifi = "Wi-Fi"
-    static let bluetooth = "Bluetooth"
-    static let mobileData = "Cотовая связь"
-    static let personalHotspot = "Режим модема"
-    static let notification = "Уведомления"
-    static let soundsHaptics = "Звуки, тактильные сигналы"
-    static let focus = "Фокусирование"
-    static let screenTime = "Экранное время"
-    static let general = "Основные"
-    static let controlCentre = "Пункт управления"
-    static let displayBrightness = "Экран и яркость"
-    static let faceIdPasscode = "Face ID и код-пароль"
-    static let battery = "Аккумулятор"
-    static let privacySecurity = "Конфиденциальность\n и безопасность"
-    static let passwords = "Пароли"
-    static let phone = "Телефон"
-    static let mail = "Почта"
-}
-
-// Image
-
-enum ConstantsImage {
-    static let user = "userIcon"
-    static let airplaneMode = "airplane"
-    static let wifi = "wifi"
-    static let bluetooth = "bluetooth"
-    static let mobileData = "antenna.radiowaves.left.and.right"
-    static let personalHotspot = "personalhotspot"
-    static let notification = "bell.badge.fill"
-    static let soundsHaptics = "speaker.wave.3.fill"
-    static let focus = "moon.fill"
-    static let screenTime = "hourglass"
-    static let general = "gear"
-    static let controlCentre = "switch.2"
-    static let displayBrightness = "textformat.size"
-    static let faceIdPasscode = "faceid"
-    static let battery = "battery.100"
-    static let privacySecurity = "hand.raised.fill"
-    static let passwords = "key.fill"
-    static let phone = "phone.fill"
-    static let mail = "envelope.fill"
-}
-
-// Color
-
-enum ConstantsColor {
-    static let systemOrange = UIColor.systemOrange
-    static let systemBlue = UIColor.systemBlue
-    static let systemGreen = UIColor.systemGreen
-    static let systemPink = UIColor.systemPink
-    static let systemRed = UIColor.systemRed
-    static let systemIndigo = UIColor.systemIndigo
-    static let systemGray = UIColor.systemGray
-}
-
