+++ SettingScreen.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Workspace
+   version = "1.0">
+   <FileRef
+      location = "self:/Users/igornikolaev/Xcode/&#x41a;&#x443;&#x440;&#x441;&#x44b; IOS/&#x41f;&#x440;&#x43e;&#x435;&#x43a;&#x442;&#x44b;/&#x414;&#x43e;&#x43c;&#x430;&#x448;&#x43d;&#x44f;&#x44f; &#x440;&#x430;&#x431;&#x43e;&#x442;&#x430;/HomeWork13/SettingScreen.xcodeproj">
+   </FileRef>
+</Workspace>
