+++ SettingScreen/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
