+++ SettingScreen/Resources/Assets.xcassets/Картинки для ДЗ/userIcon.imageset/Contents.json
