+++ SettingScreen/Sources/App/AppDelegate.swift
